requirements.txt
psycopg~=3.0.8
fastapi~=0.71.0
pydantic~=1.9.0
cryptography
astroid
Click
colorama
gunicorn
h11
isort
lazy-object-proxy
mccabe
pylint
six
starlette
uvicorn
websockets
wrapt
PyJWT~=2.3.0
stellar-sdk >= 2.0.0
requests~=2.27.1